feat: 기능 테스트 오류나지 않도록 변경
